Rename the repackager entry point and split out `Repackager`
The CLI class invoked to slim a JAR (was `burdock.Bootstrapper`) is now
`soundness.repackage`, launched with `java -cp app.jar soundness.repackage`.
It cannot be a bare default-package `repackage` — house-style SN-131 requires
every file to declare a package — so the `@main` lives in the `soundness`
umbrella file, delegating to the CLI logic `burdock.repackage()`.

The slimming logic and its shared types (`Requirement`, `Entry`, the
`Burdock-*` manifest attributes, `UserError`) move out of `Bootstrapper`
into an object renamed `Repackage` → `Repackager`, freeing the `repackage`
name and keeping the entry point a thin launcher.
